I'm adding a transitional libqt4-core.
fabo or anyone else, please revert this commit if you do not agree or you think that it's too early for this change.

btw, I'm not sure if Conflicts: libqt4-core (<< 4.4.0~beta1-1) for libqtcore4 and the like is needed in case of libqt4-core return, but I'm leaving them for now.

Now I'll start building and I'll update *.install files later tonight.