Restore QtCore ABI compatibility with binaries built with g++ 4.3 on armel.
Once Qt is rebuilt with g++ 4.4, it becomes ABI incompatible with binaries
built with g++ 4.3 on armel. That's because g++ 4.4 mangles va_list
differently on armel. As a result, affected symbols are those which have
va_list type in their argument list. Qt exports 2 such symbols:

  qvsnprintf(char *str, size_t n, const char *fmt, va_list ap);
  QString &QString::vsprintf(const char* cformat, va_list ap);

The patch uses overloading feature of the C++ language to add aliases for the
symbols above. Those aliases are mangled under g++ 4.4 in the same way as
original symbols are under g++ 4.3 and below.