Decrease failure rate in QProcess on hppa caused by #561203.
Add usleep(1000) before fork() in QProcess on hppa in order to decrease failure
rate caused by #561203. Hopefully, probablity of the failure becomes so low
that packages do not FBTFS anymore. This hack is just an ugly but very needed
workaround for a common manifestation of the bug in QProcess (patch
99_hppa_bug561203_decrease_failure_rate.diff).