Break the circular dependency between libqt4-dbus and qdbus.
By adding libqtdbus4 with the contents of libqt4-dbus in it. Then make libqt4-dbus
depend on both qdbus and libqtdbus4. The symbols file of libqtdbus4 depends
on libqt4-dbus to ensure that a dependency on libqt4-dbus ensures qdbus is
installed, also on rebuilt packages. This arrangement is transitional.
(Closes: #669278).